+++ tiger-common/src/main/java/de/gematik/test/tiger/common/exceptions/TigerPortToProcessMappingException.java
@@ -1,0 +1,12 @@
+/*
+ * ${GEMATIK_COPYRIGHT_STATEMENT}
+ */
+
+package de.gematik.test.tiger.common.exceptions;
+
+public class TigerPortToProcessMappingException extends RuntimeException {
+
+  public TigerPortToProcessMappingException(String message, Throwable throwable) {
+    super(message, throwable);
+  }
+}