+++ tiger-testenv-mgr/src/frontend/src/types/ExperimentalFeatures.ts
@@ -1,0 +1,13 @@
+/*
+ * ${GEMATIK_COPYRIGHT_STATEMENT}
+ */
+export class ExperimentalFeatures {
+  trafficVisualization: boolean = false;
+
+
+  static fromMap(map: { [x: string]: string; }): ExperimentalFeatures {
+    const features = new ExperimentalFeatures();
+    features.trafficVisualization = map["trafficvisualization"] === "true";
+    return features;
+  }
+}