+++ tiger-testenv-mgr/src/main/java/de/gematik/test/tiger/testenvmgr/controller/TigerGlobalConfigurationController.java
@@ -9,9 +9,11 @@ import de.gematik.test.tiger.common.config.TigerGlobalConfiguration;
 import de.gematik.test.tiger.testenvmgr.data.TigerConfigurationPropertyDto;
 import java.util.Comparator;
 import java.util.List;
+import java.util.Map;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -22,6 +24,7 @@ import org.springframework.web.server.ResponseStatusException;
 @RestController
 @RequestMapping("/global_configuration")
 public class TigerGlobalConfigurationController {
+
   @GetMapping()
   public List<TigerConfigurationPropertyDto> getGlobalConfiguration() {
     return TigerGlobalConfiguration.exportConfiguration().entrySet().stream()
@@ -31,6 +34,11 @@ public class TigerGlobalConfigurationController {
                     e.getKey(), e.getValue().getRight(), e.getValue().getLeft().toString()))
         .sorted(Comparator.comparing(TigerConfigurationPropertyDto::getKey))
         .toList();
+  }
+
+  @GetMapping("/{keyPrefix}")
+  public Map<String, String> loadSubsetOfConfiguration(@PathVariable String keyPrefix) {
+    return TigerGlobalConfiguration.readMap(keyPrefix);
   }
 
   @PutMapping()